requirements.txt
configparser
Flask
django-htmlmin
flask_security